src/main/java/com/martin/controller/NotificationController.java
package com.martin.controller;

import com.martin.dto.NotificationRequest;
import com.martin.entity.NotificationHistory;
import com.martin.service.NotificationService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/notifications")
public class NotificationController {

    private final NotificationService notificationService;

    public NotificationController(NotificationService notificationService) {
        this.notificationService = notificationService;
    }

    @PostMapping
    public ResponseEntity<String> sendNotification(@Valid @RequestBody NotificationRequest request) {
        try {
            notificationService.sendNotification(request);
            return ResponseEntity.status(HttpStatus.OK).body("Notification processed successfully");
        } catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error processing notification");
        }
    }

    @GetMapping("/history")
    public List<NotificationHistory> getNotificationHistory() {
        return notificationService.getNotificationHistory();
    }

}
